requirements.txt
pypdf>=5.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.3
requests>=2.31.0
numpy>=1.26.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
openai>=1.30.0
httpx<0.28
arize-phoenix>=4.26.0
opentelemetry-exporter-otlp>=1.25.0
langfuse>=2.39.0
gradio>=4.39.0
PyPDF2>=3.0.0
google-generativeai>=0.7.2
python-docx>=1.1.2
PyMuPDF>=1.24.8
python-dotenv>=1.0.1

# ADK cloud + infra
google-cloud-aiplatform>=1.68.0
google-cloud-storage>=2.18.2
google-cloud-firestore>=2.16.0
reportlab>=4.2.2
PyYAML>=6.0.2
